# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cancun Chaos Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create or login a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerBody"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players, admin]
      summary: Delete player (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /players/{id}/summary:
    get:
      operationId: getPlayerSummary
      tags: [players]
      summary: Get "while you were away" summary for player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSummary"
  /players/{id}/move:
    post:
      operationId: movePlayer
      tags: [players, map]
      summary: Move player to a new hex location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovePlayerBody"
      responses:
        "200":
          description: Move result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveResult"

  # --- CHARACTERS ---
...
  /npcs:
    get:
      operationId: listNpcs
      tags: [npcs]
      summary: List all NPCs
      responses:
        "200":
          description: NPCs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Npc"
    post:
      operationId: createNpc
      tags: [npcs, admin]
      summary: Create an NPC (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNpcBody"
      responses:
        "201":
          description: Created NPC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Npc"
  /npcs/{id}:
    patch:
      operationId: updateNpc
      tags: [npcs, admin]
      summary: Update an NPC (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNpcBody"
      responses:
        "200":
          description: Updated NPC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Npc"
    delete:
      operationId: deleteNpc
      tags: [npcs, admin]
      summary: Delete an NPC (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- ACHIEVEMENTS ---
...
  /social/gossip:
    get:
      operationId: listGossip
      tags: [social]
      summary: Get latest absurd gossip/stories
      responses:
        "200":
          description: Gossip items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GossipItem"
    post:
      operationId: generateGossip
      tags: [social]
      summary: Generate new gossip for a player action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateGossipBody"
      responses:
        "201":
          description: Generated gossip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GossipItem"
  /social/decisions/{playerId}:
    get:
      operationId: getPlayerDecisions
      tags: [social]
      summary: Get pending decisions for a player
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pending decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
    post:
      operationId: makeDecision
      tags: [social]
      summary: Make a decision
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MakeDecisionBody"
      responses:
        "200":
          description: Decision result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResult"

  # --- ADMIN ---
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        money:
          type: integer
        energy:
          type: integer
        reputation:
          type: integer
        currentHexQ:
          type: integer
        currentHexR:
          type: integer
        lastSeen:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - money
        - energy
        - reputation
        - currentHexQ
        - currentHexR
        - lastSeen
        - createdAt
    CreatePlayerBody:
      type: object
      properties:
        username:
          type: string
      required:
        - username
    UpdatePlayerBody:
      type: object
      properties:
        money:
          type: integer
        energy:
          type: integer
        reputation:
          type: integer
    PlayerSummary:
      type: object
      properties:
        playerId:
          type: integer
        moneyGained:
          type: integer
        moneyLost:
          type: integer
        eventsOccurred:
          type: array
          items:
            type: string
        visitorsCount:
          type: integer
        missionsCompleted:
          type: integer
        headline:
          type: string
      required:
        - playerId
        - moneyGained
        - moneyLost
        - eventsOccurred
        - visitorsCount
        - missionsCompleted
        - headline
    MovePlayerBody:
      type: object
      properties:
        hexQ:
          type: integer
        hexR:
          type: integer
      required:
        - hexQ
        - hexR
    MoveResult:
      type: object
      properties:
        success:
          type: boolean
        newHexQ:
          type: integer
        newHexR:
          type: integer
        energyUsed:
          type: integer
        locationName:
          type: ["string", "null"]
        eventTriggered:
          type: ["string", "null"]
      required:
        - success
        - newHexQ
        - newHexR
        - energyUsed
        - locationName
        - eventTriggered
...
    Npc:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        emoji:
          type: string
        personality:
          type: string
        dialogue:
          type: string
        locationId:
          type: ["integer", "null"]
        spawnProbability:
          type: number
        isActive:
          type: boolean
      required:
        - id
        - name
        - emoji
        - personality
        - dialogue
        - spawnProbability
        - isActive
    CreateNpcBody:
      type: object
      properties:
        name:
          type: string
        emoji:
          type: string
        personality:
          type: string
        dialogue:
          type: string
        locationId:
          type: ["integer", "null"]
        spawnProbability:
          type: number
        isActive:
          type: boolean
      required:
        - name
        - emoji
        - personality
        - dialogue
        - spawnProbability
        - isActive
...
    GossipItem:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        aboutPlayerId:
          type: ["integer", "null"]
        aboutPlayerName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - text
        - createdAt
    GenerateGossipBody:
      type: object
      properties:
        playerId:
          type: integer
        action:
          type: string
      required:
        - playerId
        - action
    Decision:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        title:
          type: string
        description:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        consequenceA:
          type: string
        consequenceB:
          type: string
        isResolved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - playerId
        - title
        - description
        - optionA
        - optionB
        - consequenceA
        - consequenceB
        - isResolved
        - createdAt
    MakeDecisionBody:
      type: object
      properties:
        decisionId:
          type: integer
        choice:
          type: string
      required:
        - decisionId
        - choice
    DecisionResult:
      type: object
      properties:
        success:
          type: boolean
        consequence:
          type: string
        moneyEffect:
          type: integer
        energyEffect:
          type: integer
        reputationEffect:
          type: integer
      required:
        - success
        - consequence
        - moneyEffect
        - energyEffect
        - reputationEffect
...